backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
rembg>=2.0.50
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
